# Optimise application lookup